requirements.txt
# ============================================================
# Mamba-Decoupled-EKF Track  —  Requirements
# ============================================================
# Core (tracking pipeline)
# ============================================================
PyYAML
numpy>=1.22.0
scipy>=1.9.0
lap
easydict
pyquaternion
shapely
opencv-python
tqdm

# ============================================================
# Deep Learning (Module A/B: Decoupled KF + Mamba)
# ============================================================
torch>=1.12.0
# mamba-ssm requires:
#   - torch >= 2.0  (for Triton kernels)
#   - CUDA toolkit matching your GPU
#   - Linux x86_64
# Install with:  pip install mamba-ssm
# If mamba-ssm is unavailable, the code auto-falls back to GRU.
mamba-ssm>=1.2.0; python_version>="3.8"

# ============================================================
# Dataset-specific (install as needed)
# ============================================================
nuscenes-devkit>=1.1.11
pandas>=1.5.0
pillow>=10.0.0
rdp

# ============================================================
# Evaluation
# ============================================================
motmetrics==1.1.3
numba>=0.58.0

# ============================================================
# Waymo (optional — requires separate setup, see README)
# ============================================================
# tensorflow>=2.6.0
# waymo-open-dataset-tf-2-11-0
# protobuf>=3.19.0
